refactor: 제목, 테이블 엘리먼트 만드는 함수 외부 모듈로 분리
